Lift the enum definition collection out of the basic block analyzer
The MIR analyzer is not the only place that discovers a type whose enums have
to be declared as datatypes, so give the collection a home on the analyzer that
owns the definitions.
